Replace audio selection.
Renamed Guide and Benefits screens. Adding comment to code to describe function/purpose. Removed useCallback for gesture handling.
